# src/main/resources/swagger/swagger.yml
swagger: "2.0"
info:
  description: "API de Gerenciamento do CBF"
  version: 4.0.0
  title: "API de Gerenciamento do CBF"
host: "sboot-cbf-api"
basePath: "/v1/cbf"
schemes:
  - "https"
paths:
  /jogadores:
    get:
      tags:
        - "sboot-cbf-api"
      operationId: "findAllJogadores"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/ListaJogadoresResponse"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
    post:
      tags:
        - "sboot-cbf-api"
      operationId: "insertJogador"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Objeto do Jogador"
          required: true
          schema:
            $ref: "#/definitions/JogadorRequest"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/StandardSuccess"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
  /jogadores/transferir:
    post:
      tags:
        - "sboot-cbf-api"
      operationId: "transferirJogador"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Objeto de Transferencia do Jogador"
          required: true
          schema:
            $ref: "#/definitions/Transferencia"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/StandardSuccess"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
  /jogadores/{id}:
    get:
      tags:
        - "sboot-cbf-api"
      operationId: "findJogadorById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Id do Jogador"
          required: true
          type: "integer"
          format: "int32"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/JogadorResponse"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
    put:
      tags:
        - "sboot-cbf-api"
      operationId: "updateJogador"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Id do Jogador"
          required: true
          type: "integer"
          format: "int32"
        - in: "body"
          name: "body"
          description: "Objeto do Jogador"
          required: true
          schema:
            $ref: "#/definitions/JogadorRequest"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/StandardSuccess"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
    delete:
      tags:
        - "sboot-cbf-api"
      operationId: "deleteJogadorById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Id do Jogador"
          required: true
          type: "integer"
          format: "int32"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/StandardSuccess"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
  /times:
    get:
      tags:
        - "sboot-cbf-api"
      operationId: "findAllTimes"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/ListaTimesResponse"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
    post:
      tags:
        - "sboot-cbf-api"
      operationId: "insertTime"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Objeto do Time"
          required: true
          schema:
            $ref: "#/definitions/TimeRequest"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/StandardSuccess"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
  /times/{id}:
    get:
      tags:
        - "sboot-cbf-api"
      operationId: "findTimeById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Id do Time"
          required: true
          type: "integer"
          format: "int32"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/TimeResponse"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
    put:
      tags:
        - "sboot-cbf-api"
      operationId: "updateTime"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Id do Time"
          required: true
          type: "integer"
          format: "int32"
        - in: "body"
          name: "body"
          description: "Objeto do Time"
          required: true
          schema:
            $ref: "#/definitions/TimeRequest"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/StandardSuccess"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
    delete:
      tags:
        - "sboot-cbf-api"
      operationId: "deleteTimeById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Id do Time"
          required: true
          type: "integer"
          format: "int32"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/StandardSuccess"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
  /times/{id}/inserirJogador:
    post:
      tags:
        - "sboot-cbf-api"
      operationId: "inserirJogadorTime"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Id do Time"
          required: true
          type: "integer"
          format: "int32"
        - in: "body"
          name: "body"
          description: "Lista de IDs de Jogadores"
          required: true
          schema:
            $ref: "#/definitions/ListaJogadoresRequest"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/StandardSuccess"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/StandardError"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/StandardError"
...
definitions:
  StandardError:
    type: "object"
    properties:
      timestamp:
        type: "integer"
        format: "int64"
      status:
        type: "integer"
      error:
        type: "string"
      message:
        type: "string"
  StandardSuccess:
    type: "object"
    properties:
      status:
        type: "integer"
      message:
        type: "string"
  JogadorRequest:
    type: "object"
    properties:
      nome:
        type: "string"
      dataNascimento:
        type: "string"
      pais:
        type: "string"
  JogadorResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      nome:
        type: "string"
      dataNascimento:
        type: "string"
      pais:
        type: "string"
      time:
        $ref: "#/definitions/TimeResponse"
  TimeRequest:
    type: "object"
    properties:
      nome:
        type: "string"
      localidade:
        type: "string"
  TimeResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      nome:
        type: "string"
      localidade:
        type: "string"
      jogadores:
        type: "array"
        items:
          $ref: "#/definitions/JogadorResponse"
  ListaJogadoresResponse:
    type: "object"
    properties:
      jogadores:
        type: "array"
        items:
          $ref: "#/definitions/JogadorResponse"
  ListaTimesResponse:
    type: "object"
    properties:
      times:
        type: "array"
        items:
          $ref: "#/definitions/TimeResponse"
  Transferencia:
    type: "object"
    properties:
      idJogador:
        type: "integer"
        format: "int32"
      idTime:
        type: "integer"
        format: "int32"
...
  ListaJogadoresRequest:
    type: "object"
    properties:
      idsJogadores:
        type: "array"
        items:
          type: "integer"
          format: "int32"
...